refactoring and test done
