+++ src/communication/network_interface_cl.h
@@ -50,6 +50,7 @@
 #include "load_common.hpp"
 #include "timezone_lib_common.h"
 #include "dynamic_array.h"
+#include "flashback.h"
 
 // forward declarations
 #if defined (SA_MODE)
@@ -444,8 +445,10 @@ extern int loaddb_destroy ();
 extern int loaddb_interrupt ();
 extern int loaddb_update_stats ();
 
+/* *INDENT-OFF* */
 extern int flashback_get_summary (dynamic_array * class_list, const char *user, time_t start_time, time_t end_time,
-				  void *summary_list, OID ** oid_list);
+				  Map_Summary &summary, OID ** oid_list);
+/* *INDENT-ON* */
 extern int flashback_get_loginfo (int trid, char *user, OID * classlist, int num_class, LOG_LSA * start_lsa,
 				  LOG_LSA * end_lsa, int *num_item, bool forward, void *info_list);
 #endif /* _NETWORK_INTERFACE_CL_H_ */
